Updated Main.lua
HUGE Rewrite of the main mouse click function! Everything should work better and be clearer! Doesn't mean all the bugs are squashed though, it's 1 am and im going to bed so some features are only halfway done, which means theres a glitch or two in there.

Added on-screen logs? IDK that's from like 4 months ago.