Updated main.lua
Phased out phases

Slimmed down flipturn()

All glitches I know of are fixed; you should be able to pick up a token and put it down without a problem, but only if you dont have tokens in the bank. There is no end condition to the game at this point.